refactored fetch into its own file
